requirements.txt
tqdm==4.64.1
transformers==4.22.2
evaluate==0.2.2
librosa==0.9.2
numpy==1.23.5
pandas==1.5.0
pyloudnorm==0.1.1
soundfile==0.11.0
info-nce-pytorch==0.1.4
InstructorEmbedding==1.0.0
sentence-transformers==2.2.2
matplotlib==3.6.1
scikit-image==0.19.3